# CsharpApiSourceCode/GeneratedCSharpAPI_NEW_2/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Authorize.Net REST API",
    "description" : "Authorize.Net REST API",
    "contact" : {
      "name" : "developer_feedback@authorize.net"
    },
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "https://api.authorize.net/"
  } ],
  "paths" : {
    "/v1/customers" : {
      "get" : {
        "tags" : [ "customers-profiles" ],
        "summary" : "Get All Customer Profiles",
        "description" : "Get a list of all customer profiles associated with this Authorize.Net account.",
        "operationId" : "getAllCustomers",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "description" : "The number of the page to return results from. For example, if limit is set to 100, and offset is set to 2, the function will return 100 transactions, beginning with the transaction that would otherwise be the transaction numbered 101 in the result set. Accepted values are 1-100000.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The number of transactions per page. Accepted values are 1-1000.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerCollection"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "read" ]
        }, {
          "basic" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "customers-profiles" ],
        "summary" : "Create A Customer Profile",
        "description" : "Create a new customer profile that contains a payment method.",
        "operationId" : "CreateCustomerAndPaymentMethod",
        "requestBody" : {
          "description" : "Contains the information required to create a customer profile with an asociated payment method.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCustomerAndPaymentMethod"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "write" ]
        }, {
          "basic" : [ ]
        } ],
        "x-codegen-request-body-name" : "CreateCustomerAndPaymentMethod"
      }
    },
...
    "/v1/notifications" : {
      "get" : {
        "tags" : [ "webhooks-notifications" ],
        "summary" : "Get Notification History",
        "description" : "Retreives the history of notifications for this account.",
        "operationId" : "getNotifications",
        "parameters" : [ {
          "name" : "deliveryStatus",
          "in" : "query",
          "description" : "Filter by delivery status. Status include the following./n/n* Delivered/n/n* Failed",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from_date",
          "in" : "query",
          "description" : "Only includes notifications after the specified date, in YYYY-MM-DD format.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "to_date",
          "in" : "query",
          "description" : "Only includes notifications before specified date, in YYYY-MM-DD format.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "The number of the page to return results from. For example, if limit is set to 100, and offset is set to 2, the function will return 100 transactions, beginning with the transaction that would otherwise be the transaction numbered 101 in the result set. Accepted values are 1-100000.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "The number of transactions per page. Accepted values are 1-1000.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetNotificationsResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "System Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "read" ]
        }, {
          "basic" : [ ]
        } ]
      }
    },
    "/v1/notifications/{notificationsId}" : {
      "get" : {
        "tags" : [ "webhooks-notifications" ],
        "summary" : "Get a Notification",
        "description" : "Retreives the details of a single notification.",
        "operationId" : "getNotification",
        "parameters" : [ {
          "name" : "notificationsId",
          "in" : "path",
          "description" : "Unique identifier of the notification. The identifier was sent in the original notification.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetNotificationResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "System Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "default" : {
            "description" : "Unexpected Error.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/errorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "oauth" : [ "read" ]
        }, {
          "basic" : [ ]
        } ]
      }
    }
  },
  "components" : {
    "schemas" : {
      "Identifier" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "reference" : {
            "type" : "string"
          }
        },
        "description" : "Model for Identifier",
        "example" : {
          "reference" : "reference",
          "email" : "email"
        }
      },
      "CustomerList" : {
        "type" : "object",
        "properties" : {
          "customers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Customer"
            }
          }
        },
        "example" : {
          "customers" : [ {
            "identifier" : {
              "reference" : "reference",
              "email" : "email"
            },
            "_links" : {
              "addresses" : null,
              "paymentMethods" : null,
              "self" : null
            },
            "description" : "description",
            "id" : "id"
          }, {
            "identifier" : {
              "reference" : "reference",
              "email" : "email"
            },
            "_links" : {
              "addresses" : null,
              "paymentMethods" : null,
              "self" : null
            },
            "description" : "description",
            "id" : "id"
          } ]
        }
      },
      "CustomerCollection" : {
        "type" : "object",
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/collectionLinks"
          },
          "_embedded" : {
            "$ref" : "#/components/schemas/CustomerList"
          },
          "totalCustomers" : {
            "type" : "integer",
            "description" : "Total number of customer profiles associated with tje Authorize.Net account that submitted the request."
          }
        },
        "example" : {
          "totalCustomers" : 0,
          "_links" : {
            "next" : null,
            "previous" : null,
            "self" : {
              "method" : "method",
              "href" : "href",
              "title" : "title"
            }
          },
          "_embedded" : {
            "customers" : [ {
              "identifier" : {
                "reference" : "reference",
                "email" : "email"
              },
              "_links" : {
                "addresses" : null,
                "paymentMethods" : null,
                "self" : null
              },
              "description" : "description",
              "id" : "id"
            }, {
              "identifier" : {
                "reference" : "reference",
                "email" : "email"
              },
              "_links" : {
                "addresses" : null,
                "paymentMethods" : null,
                "self" : null
              },
              "description" : "description",
              "id" : "id"
            } ]
          }
        }
      },
...
      "CreatePaymentMethod" : {
        "type" : "object",
        "properties" : {
          "addressId" : {
            "type" : "string",
            "description" : "Unique identifier of the address associated with the payment method."
          },
          "billTo" : {
            "$ref" : "#/components/schemas/baseAddress"
          },
          "default" : {
            "type" : "boolean",
            "description" : "Setting this to `true` sets this payment method as the default for this customer profile."
          },
          "paymentInformation" : {
            "$ref" : "#/components/schemas/paymentInformation"
          }
        },
        "description" : "Contains information about a customer's payment method."
      },
      "CreateCustomerAndPaymentMethod" : {
        "required" : [ "identifier" ],
        "type" : "object",
        "properties" : {
          "description" : {
            "type" : "string",
            "description" : "Description of the customer or customer profile. Required only when no values for `merchantCustomerId` and `email` are submitted. 255-character maximum."
          },
          "identifier" : {
            "$ref" : "#/components/schemas/Identifier"
          },
          "paymentMethod" : {
            "$ref" : "#/components/schemas/CreatePaymentMethod"
          }
        },
        "description" : "Contains the information to create a profile and an associated payment method."
      },
      "Customer" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Unique identifier of the customer profile."
          },
          "description" : {
            "type" : "string",
            "description" : "Description of the customer."
          },
          "identifier" : {
            "$ref" : "#/components/schemas/Identifier"
          },
          "_links" : {
            "$ref" : "#/components/schemas/CustomerLinks"
          }
        },
        "description" : "Response information returned after the creation of a customer profile.",
        "example" : {
          "identifier" : {
            "reference" : "reference",
            "email" : "email"
          },
          "_links" : {
            "addresses" : null,
            "paymentMethods" : null,
            "self" : null
          },
          "description" : "description",
          "id" : "id"
        }
      },
      "CustomerLinks" : {
        "type" : "object",
        "properties" : {
          "self" : {
            "$ref" : "#/components/schemas/link"
          },
          "paymentMethods" : {
            "$ref" : "#/components/schemas/link"
          },
          "addresses" : {
            "$ref" : "#/components/schemas/link"
          }
        },
        "example" : {
          "addresses" : null,
          "paymentMethods" : null,
          "self" : null
        }
      },
...
      "RetryLog" : {
        "required" : [ "logs" ],
        "type" : "object",
        "properties" : {
          "logs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Log"
            }
          }
        },
        "description" : "Logging information about notification retries."
      },
      "Notification" : {
        "required" : [ "_links", "deliveryStatus", "eventDate", "eventType", "notificationId", "webhookId" ],
        "type" : "object",
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/selfLink"
          },
          "notificationId" : {
            "type" : "string",
            "description" : "Unique identifier of the notification"
          },
          "deliveryStatus" : {
            "type" : "string",
            "description" : "Current status of the notification."
          },
          "eventType" : {
            "type" : "string",
            "description" : "Type of event that triggered the notification."
          },
          "eventDate" : {
            "type" : "string",
            "description" : "Date and time of the notifications sent."
          },
          "webhookId" : {
            "type" : "string",
            "description" : "Label of the Webhook."
          }
        },
        "description" : "Notification Model",
        "example" : {
          "webhookId" : "webhookId",
          "_links" : null,
          "notificationId" : "notificationId",
          "eventType" : "eventType",
          "deliveryStatus" : "deliveryStatus",
          "eventDate" : "eventDate"
        }
      },
      "GetNotificationResponse" : {
        "description" : "Details of a single notification.",
        "allOf" : [ {
          "$ref" : "#/components/schemas/Notification"
        }, {
          "type" : "object",
          "properties" : {
            "retryLog" : {
              "$ref" : "#/components/schemas/RetryLog"
            }
          }
        } ]
      },
      "GetNotificationsResponse" : {
        "required" : [ "_links", "notifications" ],
        "type" : "object",
        "properties" : {
          "_links" : {
            "$ref" : "#/components/schemas/selfLink"
          },
          "notifications" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Notification"
            }
          }
        },
        "description" : "List of notifications.",
        "example" : {
          "_links" : {
            "self" : {
              "method" : "method",
              "href" : "href",
              "title" : "title"
            }
          },
          "notifications" : [ {
            "webhookId" : "webhookId",
            "_links" : null,
            "notificationId" : "notificationId",
            "eventType" : "eventType",
            "deliveryStatus" : "deliveryStatus",
            "eventDate" : "eventDate"
          }, {
            "webhookId" : "webhookId",
            "_links" : null,
            "notificationId" : "notificationId",
            "eventType" : "eventType",
            "deliveryStatus" : "deliveryStatus",
            "eventDate" : "eventDate"
          } ]
        }
      },
...
      "baseAddress" : {
        "required" : [ "address1" ],
        "type" : "object",
        "properties" : {
          "address1" : {
            "type" : "string",
            "description" : "Number/letter of the address, and the street name. Required if merchant would like to use the Address Verification Service security feature. Required when using GPN Canada or Worldpay Streamline Processing Platform. <br /><br />String, up to 60 characters. Use alphanumeric characters only, without spaces, dashes, or other symbols."
          },
          "administrativeArea" : {
            "type" : "string",
            "description" : "The state or administrative area. Required only when using a European payment processor. <br /><br />String, up to 40 characters. Use alphanumeric characters only, without spaces, dashes, or other symbols. For US states, use the USPS two-character abbreviation for the state."
          },
          "company" : {
            "type" : "string",
            "description" : "Name of the purchasing company, if any. <br /><br /> String, up to 50 characters. Use alphanumeric characters only, without spaces, dashes, or other symbols."
          },
          "country" : {
            "type" : "string",
            "description" : "Country associated with the address. Required only when using a European payment processor. <br /><br />String, up to 60 characters. Use alphanumeric characters only, without spaces, dashes, or other symbols. For international payment processors, use the ISO 3166 alpha-3 code for the country."
          },
          "firstName" : {
            "type" : "string",
            "description" : "First name associated with customer's address. Required only when using a European payment processor. <br /><br />String, up to 50 characters. Use alphanumeric characters only, without spaces, dashes, or other symbols."
          },
          "lastName" : {
            "type" : "string",
            "description" : "Last name associated with customer's address. Required only when using a European payment processor. <br /><br />String, up to 50 characters. Use alphanumeric characters only, without spaces, dashes, or other symbols."
          },
          "locality" : {
            "type" : "string",
            "description" : "City or town of the address. Required only when using a European payment processor. <br /><br />String, up to 40 characters. Use alphanumeric characters only, without spaces, dashes, or other symbols."
          },
          "phoneNumber" : {
            "type" : "string",
            "description" : "Phone number associated with customer's address. <br /><br />String, up to 25 characters. For example, (123) 555-1234."
          },
          "postalCode" : {
            "type" : "string",
            "description" : "The postal or ZIP code of the customer's billing address. Required if merchant would like to use the Address Verification Service security feature. Required when using GPN Canada or Worldpay Streamline Processing Platform. String, up to 20 characters. Use alphanumeric characters only, without spaces, dashes, or other symbols."
          }
        },
        "example" : {
          "country" : "country",
          "firstName" : "firstName",
          "lastName" : "lastName",
          "phoneNumber" : "phoneNumber",
          "address1" : "address1",
          "postalCode" : "postalCode",
          "locality" : "locality",
          "company" : "company",
          "administrativeArea" : "administrativeArea"
        }
      },
...
      "errorResponse" : {
        "title" : "errorResponse",
        "required" : [ "message", "reason" ],
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Describes the error condition."
          },
          "reason" : {
            "type" : "string",
            "description" : "This is an enum field that describes the business errors returned by the server. The client can use this field to interpret errors returned by the server. "
          },
          "informationLink" : {
            "type" : "string",
            "description" : "Contains a URL link to general online documentation for this error."
          },
          "correlationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "details" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/errorDetails"
            }
          }
        },
        "description" : "Model for Error Response"
      },
      "errorDetails" : {
        "title" : "errorDetails",
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "description" : "Describes the error condition."
          },
          "location" : {
            "type" : "string",
            "description" : "Indicates the name of a parameter."
          }
        },
        "description" : "Error details object."
      },
...
      "collectionLinks" : {
        "type" : "object",
        "properties" : {
          "self" : {
            "$ref" : "#/components/schemas/link"
          },
          "next" : {
            "$ref" : "#/components/schemas/link"
          },
          "previous" : {
            "$ref" : "#/components/schemas/link"
          }
        },
        "example" : {
          "next" : null,
          "previous" : null,
          "self" : {
            "method" : "method",
            "href" : "href",
            "title" : "title"
          }
        }
      },
      "link" : {
        "type" : "object",
        "properties" : {
          "href" : {
            "type" : "string",
            "description" : "URI of the linked resource."
          },
          "title" : {
            "type" : "string",
            "description" : "Label of the linked resource."
          },
          "method" : {
            "type" : "string",
            "description" : "HTTP method of the linked resource."
          }
        },
        "example" : {
          "method" : "method",
          "href" : "href",
          "title" : "title"
        }
      },
...
      "selfLink" : {
        "type" : "object",
        "properties" : {
          "self" : {
            "$ref" : "#/components/schemas/link"
          }
        },
        "example" : {
          "self" : {
            "method" : "method",
            "href" : "href",
            "title" : "title"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "basic" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "oauth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://account.authorize.net/oauth/authorize",
            "scopes" : {
              "read" : "List and view details",
              "write" : "Create, list, and view details"
            }
          }
        }
      }
    }
  },
...
}